+++ src/components/container.rs
@@ -1,0 +1,8 @@
+use yew::prelude::*;
+#[function_component]
+pub fn Container() -> Html {
+    html! {
+        <div id="container">
+        </div>
+    }
+}
