+++ src/components/footer.rs
@@ -1,0 +1,8 @@
+use yew::prelude::*;
+
+#[function_component]
+pub fn Footer() -> Html {
+    html! {
+        <div id="footer">{"footer"}</div>
+    }
+}
