+++ src/components/mod.rs
@@ -1,0 +1,3 @@
+pub mod container;
+pub mod footer;
+pub mod header;
